Overhaul of RCC, GPIO, Serial. Add SPI, Delay
* Working support for RCC configuration, with example
* Working support for SPI, with example
* Working support for Serial, with example
* Working support for Delay, with example
* Add memory.x
* Add rustfmt.toml
* Document codegen options in Cargo.toml
* Added myself as an author

The RCC peripheral returns a "Core Clock Distribution and Reset"
structure, which represents that the core clock configuration is
frozen. However we retain the right to switch most PKSUs on the fly,
to fine-tune PLL frequencies, and to enable / reset peripherals.

For now, use a local build of the stm32h7 crate. This keeps us
up-to-date without Adam having to issue patch releases all the
time. Hopefully this will stabilise and can be reverted soon. This is
documented in the "hacking" section of the README

Tested on STM32H743 Rev Y

stm32h7xx-hal PR #2

See PR #1 for context