Added pwm channel info to pwm::Pins
Before, it was possible to pass any tuple of TIMx pins to TIMx.pwm(...)
and it would succeed: you could pass two different pins for CH1 or pins
for CH3 and CH4 and either way you would get PWM channels 1 and 2
which would end up writing to the wrong peripheral registers and trying
to use PWM channels with no pins configured.

To fix this, pwm::Pins<TIM> was changed to pwm::Pins<TIM, CHANNEL> and
a pin is now associated with a specific PWM channel.

Tuples of pins and channels are now marked with which channels they
are associated, such as (C1, C2, C3) or (C4, C2)

Pins is only implemented for tuples of channels with no repeated
channels, and each Pins implementation uses the correct Channels tuple
based on the channels for the tuples of pins it is implemented for.

The implementation of Pins<TIM, CHANNEL> for tuples of pins is now
provided by the macro pins_tuples! and includes every valid pin
combination in any possible order (60 total: 6 combinations of 2
channels with 2 orders each, 4 combinations of 3 channels with 6
orders each, and 1 combination of 4 channels with 24 orders)

A tuple of pins passed to pwm(...) now is checked to make sure that
the right pins are associated with the
right channels.