Extend i2c::Write implementation method until transaction finishes
Currently the method returns once the final byte is written to the FIFO, not
when the transaction finishes. Although this saves one byte's of time, it is
somewhat awkward because:

* Calling another i2c method during this time will fail.
* I2C events are not raised until the transfer ends.

At some point we will need to re-write all of this for async and DMA, but for
now I think this is less confusing.