Improvements to serial
* Allow baud rate in Hertz/KiloHertz as well as Bps
* Accept a baud rate instead of a configuration struct. In the case that a baud
rate is specified, the defaults for the other parameters (8N1) are used.
* Rename `usart` / `usart_unchecked` extension methods to `serial` /
`serial_unchecked`. The method names elsewhere match the module name, and also
some peripherals are `uart` not `usart`, so I think this is an improvement. Keep
the old methods for backwards compatibility.

No breakage, but I have updated the example.