Add DMBs before and after the stream read/writes normal memory
Without barrier instructions, the Cortex-M7 core can reorder the execution of
transfers to normal and device memory with respect to each other.

This does not correspond to the compiler_fence calls, which are concerned with
generating the correct _program order_. The Cortex-M7 core does not reorder the
transfers to device memory (the DMA configuration) with respect to each other.