Implement DMA for constant source buffers
The lack of support for constant source buffers was first raised by
@ryan-summers here
https://github.com/stm32-rs/stm32h7xx-hal/pull/153#discussion_r515865869

This solution adds a 5th generic type parameter to `Transfer`, and uses this to
implement transfer for two different type constraints on `BUF`.

The additional generic type parameter will also be useful for MDMA support,
which will also need its own versions of `init`, `next_transfer`,
`next_transfer_with` and so on.

The downside is the additional complexity of another type parameter