Finish support for 2nd DAC on rm0455 parts
* DAC peripheral on rm0433/rm0399 parts is still called "DAC12" to match
reference manual naming. We keep this, even though it is now quite confusing
considering the existance of two separate DACs on rm0455 parts
* Complete the HAL upgrade to support two DACs. We need to add the peripheral as
a generic type on the Channel type, because we need different codegen for Cx on
DAC2 compared Cx on DAC1. Add the REC type as an associated type for DacExt
implementations.
* 2nd DAC peripheral on rm0455 parts is actually capable of "autonomous mode"
operation (srd/d3), but the D3AMR/SRDAMR register still needs some work in
stm32-rs