Ethernet: Remove extraneous interrupt-free regions
There is no gain from disabling interrupts between the memory barriers
and writing to the DMA engine descriptor ring pointers - even if an
interrupt is executed between the barrier and the atomic store to the
pointer, it doesn't make the descriptors any less valid.

Interrupts are also irrelevant for transmit/receive ring initialisation,
which needs to run wiht the DMA engine disabled anyway.

GitHub: Fixes #197.