dma: add clear_transfer_complete_flag
Currently there is 2 PCLK delay (equivalent to up to 8 CPU cycles) when
clearing transfer complete flags for the update to trickle through the
bus matrices, the peripheral and the irq synchronizer after clearing an
interrupt flag.

In many cases the delay is not required because there are enough instruction
following before exiting an ISR that would fully absorb the delay.

This change adds a method to clear the flag without additional delay.