dma: add next_dbm_transfer_with
This add a new unsafe method to the DMA API for significantly faster
transfer handling.

In many cases the safety and universality offered by `next_transfer_with`
is not necessary and costly in terms of cycles. This is especially the case
when multiple double-buffer DMA streams are handled together and when the
risk of loosing the race to the inactive buffer against the running DMA
transfer is either acceptable or excluded by design of the handler.