Expose PinMode as public
With the new GPIO API it is not possible anymore to create a generic
functions that would turn Pin into PushPull unless the current mode of
the Pin is hardcoded.

This used to work:

 pub struct LedUser(pub gpio::gpioc::PC7<gpio::Output<gpio::PushPull>>);

 impl LedUser {
     pub fn new<Mode>(pin: gpio::gpioc::PC7<Mode>) -> Self {
         LedUser(pin.into_push_pull_output())
     }
 }

I would like to set <Mode: PinMode>, but for that, PinMode needs to be
made public.

Signed-off-by: Petr Horacek <hrck@protonmail.com>