Convert ResetMode to own witness pin and track past pin mode
Previously when resetting it would assume the pin was in the ORIG pin
mode and try to convert it to the ORIG pin mode, which the if
statements in mode blocked. Now by tracking the witness pin and the
original MODE the reset can properly set the mode register bits.