Add halfduplex to serial::config::Config and example
This allows the USART's to be used with serial devices that operate in a half-duplex mode (transmit and receive on one wire). An example was also added (serial-halfduplex) and has been tested with a NUCLEO-H723ZG board. It has also been tested with a Frysky RX4R RC receiver Smart Port interface.