Remove references to `static mut`
rust/#114447 made `&mut MY_STATIC_MUT` a warning, with the intention of
disallowing it entirely in the 2024 edition. We never used this in the crate
itself, but it is used in many examples.

This commit removes these references from the examples. In general the new
approach is to change `static mut` of type `X` to type `MaybeUninit<X>`. These
can then cleanly be accessed with
[`MY_STATIC_MUT.write()`](https://doc.rust-lang.org/core/mem/union.MaybeUninit.html#method.write)
followed by
[`MY_STATIC_MUT.assume_init_mut()`](https://doc.rust-lang.org/core/mem/union.MaybeUninit.html#method.assume_init_mut).

Where we need to cast a `MaybeUninit<[X; N]>` to a `[MaybeUninit<X>; N]` in
order to initialise it element-by-element, the
`&mut *ptr::addr_of_mut!(MY_STATIC_MUT)` construction is still used. After
initialisation, we can then access it with `MY_STATIC_MUT.assume_init_mut()`.

In a few DMA Examples we were taking a reference `&mut TARGET_BUFFER` without
initialising TARGET_BUFFER. This commit changes this to
`TARGET_BUFFER.assume_init_mut()` *without* initialising `TARGET_BUFFER`. This
was and remains UB.

* USB Examples: `static mut EP_MEMORY` now has type `MaybeUninit<[u32; 1024]>`
* Ethernet Examples: `static mut DES_RING` now has type `MaybeUninit<ethernet::DesRing<4, 4>>`